% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_data.R
\name{beautify_dir}
\alias{beautify_dir}
\title{beautify files in directory}
\usage{
beautify_dir(path, pattern = NULL, beautifier = beautify_file)
}
\arguments{
\item{path}{directory}

\item{pattern}{an optional regular expression. Only file
names which match the regular expression will be beautified.}

\item{beautifier}{function to use to format data}
}
\description{
beautify file in a directory.
}
\examples{
#' # beautify_dir ----

# simulate a directory with many files ----
f_json <- system.file(package = "prettyfile", "data_examples/simple.json")
f_xml <- system.file(package = "prettyfile", "data_examples/simple.xml")
f_css <- system.file(package = "prettyfile", "data_examples/simple.css")
f_sql <- system.file(package = "prettyfile", "data_examples/simple.sql")
dir.create("data_files")
file.copy(c(f_json, f_xml, f_css, f_sql), to = "data_files" )

beautify_dir("data_files")
}
