% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_data.R
\name{beautify_str}
\alias{beautify_str}
\alias{beautify_current_doc}
\title{beautify a string}
\usage{
beautify_str(str, format = "json")

beautify_current_doc()
}
\arguments{
\item{str}{json, xml, css or sql string}

\item{format}{format data contained in the string (one of json, xml, css or sql)}
}
\description{
beautify a string containing either json, xml, css or sql content.

\code{beautify_current_doc} is not to be used
by the user.
}
