% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minify.R
\name{minify_str}
\alias{minify_str}
\title{minify a string}
\usage{
minify_str(str, format = "json")
}
\arguments{
\item{str}{json, xml, css or sql string}

\item{format}{format data contained in the string (one of json, xml, css or sql)}
}
\description{
minify a string containing either json, xml, css or sql content.
}
